requirements.txt
# NOTE: torch is installed separately by scripts/install_torch.sh to prefer a CUDA wheel when available.
accelerate==1.12.0
bert-score==0.3.13
datasets==3.6.0
evaluate==0.4.6
numpy==2.1.3
packaging==24.2
peft==0.18.1
pyyaml==6.0.2
rouge-score==0.1.2
scipy==1.15.1
sentencepiece==0.2.1
tokenizers==0.22.0
tqdm==4.67.1
transformers==4.57.6
